feat(weather): WIP fixed sunrise and sunset Double to convert easily to date

+++ LeBaluchon/Models/Weather/Weather.swift
@@ -23,8 +23,8 @@ struct WeatherType: Codable, Hashable {
 
 struct WheatherSys: Codable, Hashable {
     let country: String
-    let sunrise: Int
-    let sunset: Int
+    let sunrise: Double
+    let sunset: Double
 }
 
 struct WheatherMain: Codable, Hashable  {
